fix: update JIT load/store instructions to use 32-bit offsets

+++ PolkaVM/Sources/CppHelper/jit_instructions.hh
@@ -28,13 +28,13 @@ bool jit_emit_load_imm_jump(void* _Nonnull assembler, const char* _Nonnull targe
 bool jit_emit_load_imm_jump_ind(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t src_reg, uint32_t value, uint32_t offset);
 
 // Load from memory instructions
-bool jit_emit_load_u8(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int16_t offset);
-bool jit_emit_load_i8(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int16_t offset);
-bool jit_emit_load_u16(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int16_t offset);
-bool jit_emit_load_i16(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int16_t offset);
-bool jit_emit_load_u32(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int16_t offset);
-bool jit_emit_load_i32(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int16_t offset);
-bool jit_emit_load_u64(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int16_t offset);
+bool jit_emit_load_u8(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int32_t offset);
+bool jit_emit_load_i8(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int32_t offset);
+bool jit_emit_load_u16(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int32_t offset);
+bool jit_emit_load_i16(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int32_t offset);
+bool jit_emit_load_u32(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int32_t offset);
+bool jit_emit_load_i32(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int32_t offset);
+bool jit_emit_load_u64(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int32_t offset);
 bool jit_emit_load_reserved(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg);
 
 // Load from memory instructions (direct addressing)
@@ -47,10 +47,10 @@ bool jit_emit_load_i32_direct(void* _Nonnull assembler, const char* _Nonnull tar
 bool jit_emit_load_u64_direct(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint32_t address);
 
 // Store instructions (register to memory)
-bool jit_emit_store_8(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t ptr_reg, uint8_t src_reg, int16_t offset);
-bool jit_emit_store_16(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t ptr_reg, uint8_t src_reg, int16_t offset);
-bool jit_emit_store_32(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t ptr_reg, uint8_t src_reg, int16_t offset);
-bool jit_emit_store_64(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t ptr_reg, uint8_t src_reg, int16_t offset);
+bool jit_emit_store_8(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t ptr_reg, uint8_t src_reg, int32_t offset);
+bool jit_emit_store_16(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t ptr_reg, uint8_t src_reg, int32_t offset);
+bool jit_emit_store_32(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t ptr_reg, uint8_t src_reg, int32_t offset);
+bool jit_emit_store_64(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t ptr_reg, uint8_t src_reg, int32_t offset);
 bool jit_emit_store_u8(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t ptr_reg, int16_t offset, uint8_t src_reg);
 bool jit_emit_store_u16(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t ptr_reg, int16_t offset, uint8_t src_reg);
 bool jit_emit_store_u32(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t ptr_reg, int16_t offset, uint8_t src_reg);
@@ -232,7 +232,7 @@ bool jit_emit_memcpy(void* _Nonnull assembler, const char* _Nonnull target_arch,
 bool jit_emit_memset(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t value_reg, uint8_t count_reg);
 
 // Load-Effective-Address
-bool jit_emit_lea(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int16_t offset);
+bool jit_emit_lea(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t ptr_reg, int32_t offset);
 
 // Register operations
 bool jit_emit_copy(void* _Nonnull assembler, const char* _Nonnull target_arch, uint8_t dest_reg, uint8_t src_reg);
